Extract withdrawnJson.ts
